+++ service_impl/src/booking.rs
@@ -125,7 +125,7 @@ where
 
     async fn get_for_week(
         &self,
-        calendar_week: i32,
+        calendar_week: u8,
         year: u32,
         context: Authentication<Self::Context>,
     ) -> Result<Arc<[Booking]>, ServiceError> {
